requirements.txt
Django==3.2.13
djongo==1.3.6
pymongo==3.12.3
sqlparse==0.2.4
asgiref==3.3.4
djangorestframework==3.12.4
djangorestframework_simplejwt==4.8.0
gunicorn==20.1.0
PyJWT==2.1.0
python-dotenv==0.19.2
pytz==2021.3
tzdata==2021.5
whitenoise==5.3.0
requests
Pillow
pymongo[srv]